+++ docs/api_ref/index.rst
@@ -5,7 +5,7 @@ This section gives an overview of all public ``waveresponse`` objects,
 functions and methods.
 
 
-.. currentmodule:: waveresponse
+.. currentmodule:: waveresponse2
