+++ src/waveresponse2/_standardized1d.py
